# Conflicts:
#	projects/dom/src/lib/dom.item.ts

+++ projects/dom/src/lib/dom.component.ts
@@ -728,6 +728,9 @@ export class DomComponent extends LoadingComponent<any, any> implements OnInit, 
   }
 
   protected _loadData() {
+    if (!this._accountId || !this._instrument)
+      return;
+
     this._loadPositions();
     this._loadOrderBook();
   }
@@ -911,7 +914,7 @@ export class DomComponent extends LoadingComponent<any, any> implements OnInit, 
     if (trade.instrument?.symbol !== this.instrument?.symbol) return;
 
     const item = this._getItem(trade.price);
-    // console.log('_handleQuote', trade.price, trade.volume);
+    console.log('_handleQuote',trade.updateType, trade.price, trade.volume);
     this._handleMaxChange(item.handleQuote(trade), item);
 
     this.detectChanges();
@@ -1152,6 +1155,9 @@ export class DomComponent extends LoadingComponent<any, any> implements OnInit, 
   }
 
   private _cancelOrderByClick(column: string, item: DomItem) {
+    if (!item[column]?.canCancelOrder)
+      return;
+
     if (item.orders?.orders?.length)
       this._ordersRepository.deleteMany(item.orders.orders).subscribe(
         () => console.log('delete order'),
